+++ src/Ext/TechnoType/Hooks.cpp
@@ -116,6 +116,17 @@ DEFINE_HOOK(0x711FDF, TechnoTypeClass_RefundAmount_FactoryPlant, 0x8)
 	return 0;
 }
 
+DEFINE_HOOK(0x51A002, InfantryClass_UpdatePosition_BeforeInfiltrate, 6)
+{
+	GET(InfantryClass*, pSpy, ESI);
+	GET(BuildingClass*, pBuilding, EDI);
+
+	auto const pSpyExt = TechnoTypeExt::ExtMap.Find(pSpy->Type);
+	pSpyExt->WhenInfiltratesInto(pSpy, pBuilding);
+
+	return 0;
+}
+
 
 DEFINE_HOOK(0x71464A, TechnoTypeClass_ReadINI_Speed, 0x7)
 {