+++ docs/Whats-New.md
@@ -470,6 +470,7 @@ New:
 - Allow customizing charge turret delays per burst on a weapon (by Starkku)
 - Unit `Speed` setting now accepts floating point values (by Starkku)
 - Crush expanded: crusher level system, weapon detonation events when unit crushes something or gets crushed (by Aephiex)
+- Weapon detonation events when unit infiltrates (by Aephiex)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)